Move optimize script into scripts dir
